agent/backend/pyproject.toml
[tool.poetry]
name = "university-agent-backend"
version = "0.1.0"
description = "University Agent backend project"
authors = ["Yusuf Baykaloğlu <yusuff.baykaloglu@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = ">=0.104.1"
uvicorn = ">=0.24.0"
pydantic = ">=2.4.2"
datasets = ">=2.14.5"
langchain = ">=0.0.325"
langchain-community = ">=0.0.10"
sentence-transformers = ">=2.2.2"
numpy = ">=1.24.3"
python-dotenv = ">=1.0.0"
smolagents = ">=0.0.5"
rank_bm25 = "*"
duckduckgo-search = "*"
beautifulsoup4 = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"